// src/app/shared/model/IReuest.ts
export interface IRequest {
    customerName: string
    branchName: string
    branchAddress: string
    lat: string
    lang: string
    comment: string
    speedInternet: number
    speedInternetType: string
    speedVPN: number
    speedVPNType: string
    speedWifi: number
    speedWifiType: string
    priCount: number
    // dateFinished: string
    // dateSendToVendor: string
    localContactName: string
    localContactMobile1: string
    localContactMobile2: string
    localContactEmail: string
    // userId: number
    // vendorId: number
    // statusId: number
  }
  